requirements.txt
flask
requests
python-dotenv
openai
chromadb
picamera2
opencv-python
celery
autogen
